Wiki citation support
Added a citation builder for the Wiki format. The CitationBuilder trait can be implemented by other formats we wish to support. We can consider whether the prefixes handed to the add() and try_add() functions of the builder should instead be enum values so that the Reference type doesn't have to know about the low-level markup for the citation formats.