Wrapper enum for Attribute and encapsulation of types
Previously, an Attribute was a struct containing its value in String
representation and a priority. It was found to be severely limiting that
Attribute could only represent its data as String, and has therefore
been changed to an Enum containing all possible AttributeTypes
(renamed from InternalAttributeKey). This change will make it easier to
support SchemaOrg in the near future.

Storing priority alongside the Attribute always felt wrong. A new type,
StoredAttribute, is now used for AttributeCollection in its internal
representation.

Many types and functions have been moved to different files to better
comply with encapsulation issues.

A small testcase has been added in tests/test.rs. The plan is to
massively expand the testbase with several HTML examples.

Co-authored-by: BX326 <44841660+BX326@users.noreply.github.com>