Simple BibTeX support
A Reference now has a private and generic build_citation() function which takes a CitationBuilder. The functions for returning a citation in a specific format simply calls this generic function with a CitationBuilder instance corresponding to the format. Additionally, the citation module has been moved to its own file. The next step is to support APA and handle author and date attributes across the citation formats.