Significant refactor
Builder patterns for AttributeCollection, AttributeConfig and GenerationOptions. Currently, it's awkward to deal with GenerationOptions and by extension the attribute priority configuration. This work is an attempt to alleviate this problem by utilizing the builder pattern and appropriate defaults. It also simplifies the AttributeCollection structure; the StoredAttribute type has become redundant and priority handling is now implicit. Now, each attribute is parsed individually (instead of all attributes at once) according to our sorted priority lists. When a metadata parser returns a hit for an attribute, we can simply use that value without parsing further. This results in a performance improvement as we no longer parse exhaustively.

Note: While the application builds with these changes, there's still some work to be done. For instance, AttributeParser::parse_attribute() has yet to be implemented for the OpenGraph and Schema.org parsers.