CLI: Naive priority handling
Now, the user can specify a single prioritized metadata format when using the CLI. However, the implementation currently has the unintended consequence that the specified format is searched exclusively rather than being just prioritized. I tried to re-implement AttributePriority to contain fields such as first and second to reflect the priorities, but this led to some challenges. Firstly, these fields should probably be convertible to a priority list. Second, there's the issue of how to initialize the other fields if only a single field is set. If using static defaults, there's the risk that another metadata format may be overwritten when the user calls a setter function rather than just dropping in priority. We have to put some thought into reworking AttributePriority.