Finish DOI parsing and rework Date attribute
This commit adds the remaining attributes (to the degree which is possible)
to the DOI parser. Tests for the DOI parser remain an issue, as it is uncertain
how to handle them without an internet connection.

There is also a rework of the Date attribute. The Date attribute is now
represented by a date enum which can represent complete and partially complete
dates. Expected results in tests needed to change in order to indicate which
type of date they represent.