MVP for web app
This commit adds minimal functionality to the web app.
There is a very simple interface, that given an URL
and a citation format, will output the given citation.
There is still a good amount of work to be done. The UI
design is lacklustre and does not support all the features of our
backend.