Replace archive checkbox with a toggle button
The archive checkbox has been replaced with a toggle button.
It has also been aligned correctly with the translation
options. The inner label of the button is replaced
with a checkmark or a cross depending on button state.

There are potential issues with visibility on the off state,
but I think this is good for now.